refactor(dispatch): extract the state/mutex layer to dispatch-state.ts
Phase B step 1: move the .dispatch-state.json layer out of dispatch.ts into a leaf module —
the stateQueue async mutex + withStateLock, load/save/getStateFile, dispatch-marker helpers,
getFailureRecord/cooldownForFailure, trimDispatched, clearDispatchMarker. dispatch.ts imports
them back and re-exports loadDispatchState + clearDispatchMarker (public surface).

This isolates the dispatch-state mutex into ONE module — structurally guaranteeing the
single-instance property the whole fire path depends on (a split stateQueue would silently
break the lock → duplicate fires). Pure relocation; saveDispatchState kept byte-identical
(the atomic temp-then-rename improvement is noted as a separate follow-up). The file
path/dispatchSeq shape (a cross-package ledger contract) is unchanged.

dispatch.ts: 1,652 → 1,588. Verified: typecheck + lint clean; dispatch/continuation/budget/
task-service/restart-recovery 104/0. (Full suite + binary + docker dispatch-smoke run after
the remaining fire-core modules land.)
